Experimental rkt attach for stdio;  preserve cwd.
A fairly major change, introducing rkt-run-slave, which runs
in the container before the target program.  It can be asked to
wait until notification that rkt attach has run on the host, before
executing the target.

Currently, rkt attach must be run externally to rkt-run, but this will
be added imminently.

Also, rkt-run-slave can chdir to desired cwd.

Specifying extra environment in the config file is removed, as
no longer required.