Config file: incompatible change - options now split by mode.
Allows to specify different options to be used in interactive and
batch modes, or options common to both.  This anticipates wanting
to use stage1-fly in batch mode, as a work-around for the bugs in
the experimental rkt attach.  However, stage1-fly is not suitable
for interactive mode, due to the lack of tty support.